requirements.txt
Flask
qrcode
qrcode[pil]
Pillow
pyzbar
gunicorn
opencv-python-headless
gunicorn



